+++ NEWS.md
@@ -1,4 +1,6 @@
 # 3.0.0
+* Removed arguments `max.level`, `use.forward.search` in exported function `PresenceAssignActivity`.
+
 * Removed redundant arguments `sig.pres.test.nbinom.size`, `sig.pres.test.p.thresh`, `sig.pres.test.q.thresh`,  in various internal and exported
 functions.